+++ src/index.ts
@@ -4,6 +4,7 @@ export * from "./lib/date";
 export * from "./lib/enum";
 export * from "./lib/guid";
 export * from "./lib/localStorage";
+export * from "./lib/mimeType";
 export * from "./lib/number";
 export * from "./lib/object";
 export * from "./lib/string";