+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `isImageMimeType`, `isPdfMimeType`, `isVideoMimeType`, `isAudioMimeType`, `isTextMimeType`, and `isApplicationMimeType` MIME type utility functions
+
 ## [1.4.0] - 2025-06-24
 
 ### Added